handle keyvault error
