+++ src/common/utils.ts
@@ -30,3 +30,23 @@ export function shuffleList<T>(array: T[]): T[] {
     }
     return array;
 }
+
+export function getValidUrl(endpoint: string): URL {
+    try {
+        return new URL(endpoint);
+    } catch (error) {
+        if (error.code === "ERR_INVALID_URL") {
+            throw new RangeError("Invalid endpoint URL.", { cause: error });
+        } else {
+            throw error;
+        }
+    }
+}
+
+export function getUrlHost(url: string) {
+    return new URL(url).host;
+}
+
+export function instanceOfTokenCredential(obj: unknown) {
+    return obj && typeof obj === "object" && "getToken" in obj && typeof obj.getToken === "function";
+}
