+++ src/keyvault/AzureKeyVaultKeyValueAdapter.ts
@@ -4,8 +4,8 @@
 import { ConfigurationSetting, isSecretReference, parseSecretReference } from "@azure/app-configuration";
 import { IKeyValueAdapter } from "../IKeyValueAdapter.js";
 import { KeyVaultOptions } from "./KeyVaultOptions.js";
+import { getUrlHost } from "../common/utils.js";
 import { SecretClient, parseKeyVaultSecretIdentifier } from "@azure/keyvault-secrets";
-import { OperationError } from "../error.js";
 
 export class AzureKeyVaultKeyValueAdapter implements IKeyValueAdapter {
     /**
@@ -25,7 +25,7 @@ export class AzureKeyVaultKeyValueAdapter implements IKeyValueAdapter {
     async processKeyValue(setting: ConfigurationSetting): Promise<[string, unknown]> {
         // TODO: cache results to save requests.
         if (!this.#keyVaultOptions) {
-            throw new OperationError("Failed to process the key vault reference. The keyVaultOptions is not configured.");
+            throw new RangeError("Failed to process the key vault reference. The keyVaultOptions is not configured.");
         }
 
         // precedence: secret clients > credential > secret resolver
@@ -35,7 +35,7 @@ export class AzureKeyVaultKeyValueAdapter implements IKeyValueAdapter {
 
         const client = this.#getSecretClient(new URL(vaultUrl));
         if (client) {
-            // TODO: what if error occurs when reading a key vault value? Now it breaks the whole load.
+            // If the credential of the secret client is wrong, AuthenticationError will be thrown.
             const secret = await client.getSecret(secretName, { version });
             return [setting.key, secret.value];
         }
@@ -44,14 +44,21 @@ export class AzureKeyVaultKeyValueAdapter implements IKeyValueAdapter {
             return [setting.key, await this.#keyVaultOptions.secretResolver(new URL(sourceId))];
         }
 
-        throw new OperationError("Failed to process the key vault reference. No key vault credential or secret resolver callback configured, and no matching secret client could be found.");
+        // When code reaches here, it means the key vault secret reference is not resolved.
+
+        throw new RangeError("Failed to process the key vault reference. No key vault credential or secret resolver callback is configured.");
     }
 
+    /**
+     * 
+     * @param vaultUrl - The url of the key vault.
+     * @returns 
+     */
     #getSecretClient(vaultUrl: URL): SecretClient | undefined {
         if (this.#secretClients === undefined) {
             this.#secretClients = new Map();
-            for (const c of this.#keyVaultOptions?.secretClients ?? []) {
-                this.#secretClients.set(getHost(c.vaultUrl), c);
+            for (const client of this.#keyVaultOptions?.secretClients ?? []) {
+                this.#secretClients.set(getUrlHost(client.vaultUrl), client);
             }
         }
 
@@ -69,8 +76,4 @@ export class AzureKeyVaultKeyValueAdapter implements IKeyValueAdapter {
 
         return undefined;
     }
-}
-
-function getHost(url: string) {
-    return new URL(url).host;
 }
